requirements.txt
bitarray==2.3.4
bloompy==0.1.1
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mmh3==3.0.0
pybase62==0.4.3
pymongo==3.12.1
Werkzeug==2.0.2
